requirements.txt
Flask
python-docx
PyMuPDF
pytesseract
spacy
pdf2image
Pillow
